requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
openai-whisper==20231117
TTS==0.22.0
ollama==0.1.6
PyMuPDF==1.23.26
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1 